packages/sinapsis_data_writers/pyproject.toml
[project]
name = "sinapsis-data-writers"
version = "0.1.15"
description = "Templates to write DataContainer content to file."
authors = [{ name = "SinapsisAI", email = "dev@sinapsis.tech" }]

readme = "README.md"
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
    "sinapsis>=0.1.1",
    "sinapsis-generic-data-tools",
]

[project.optional-dependencies]
pydub = ["pydub>=0.25.1"]
soundfile = ["soundfile>=0.12.1"]
pdf-image-writer = ["pdf2image>=1.17.0"]
all = ["sinapsis-data-writers[pdf-image-writer,pydub,soundfile,opencv,ffmpeg]"]
opencv = [
    "opencv-python-headless>=4.11.0.86",
]
ffmpeg = [
    "ffmpeg-python>=0.2.0",
]

[tool.uv.sources]
sinapsis-data-writers = { workspace = true }
sinapsis-generic-data-tools = { workspace = true }
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"



[[tool.uv.index]]
name = "sinapsis"
url = "https://pypi.sinapsis.tech/"


[project.urls]
Homepage = "https://sinapsis.tech"
Documentation = "https://docs.sinapsis.tech/docs"
Tutorials = "https://docs.sinapsis.tech/tutorials"
Repository = "https://github.com/Sinapsis-AI/sinapsis-data-tools.git"
